assignment 42 complete

+++ app/controllers/comments_controller.rb
@@ -4,29 +4,60 @@ class CommentsController < ApplicationController
   before_action :authorize_user, only: [:destroy]
 
   def create
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
+    # common code goes here
+    if params[:post_id]
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
+      comment.user = current_user
 
-    if comment.save
-      flash[:notice] = "Comment saved successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment failed to save"
-      redirect_to [@post.topic, @post]
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save"
+        redirect_to [@post.topic, @post]
+      end
+
+    else params[:topic_id]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment failed to save"
+        redirect_to [@topic]
+      end
+
     end
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
+    if params[:post_id]
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.find(params[:id])
 
-    if comment.destroy
-      flash[:notice] = "Comment was deleted successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post.topic, @post]
+      if comment.destroy
+        flash[:notice] = "Comment was deleted successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@post.topic, @post]
+      end
+
+    else params[:topic_id]
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.find(params[:id])
+
+      if comment.destroy
+        flash[:notice] = "Comment was deleted successfully."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@topic]
+      end
     end
   end
 
@@ -40,7 +71,11 @@ class CommentsController < ApplicationController
     comment = Comment.find(params[:id])
     unless current_user = comment.user || current_user.admin?
       flash[:alert] = "You do not have permission to delete a comment."
-      redirect_to [comment.post.topic, comment.post]
+      if params[:post_id]
+        redirect_to [comment.post.topic, comment.post]
+      else params[:topic_id]
+        redirect_to [comment.topic]
+      end
     end
   end
 end
